Update and rename cipher.py to cube_rotator.py
Re-worked get_matrix(key_s_complete) to accept an argument, which is a string containing all the letters of the alphabet in the order they should be placed on the cube according to the separate cryptosystem documentation. 

Re-worked main() to be a function scramble_cube(matrix, k_r) which scrambles matrix using the rotation key k_r.